Particle System Project
My concept is to create a particle system that could bring satisfying to the viewer. When you start to run the file, each pixel(rect) will begin moving around and as the times goes the pattern will stop to a random wallpaper, my next step is to make each pixel change throughout the time and form a digital clock on the page.